% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_exponential_families.R
\name{check_full_rank}
\alias{check_full_rank}
\title{Check Full Rank Condition for Exponential Family}
\usage{
check_full_rank(ef, theta_range = seq(0.1, 5, by = 0.1))
}
\arguments{
\item{ef}{An "expfam" object}

\item{theta_range}{Numeric vector of theta values to evaluate over}
}
\value{
Logical: TRUE if the family appears to be full rank
}
\description{
Verifies whether an exponential family is of full rank by checking that
eta(theta) is not constant over the natural parameter space.
}
\examples{
ef <- exponential_family("normal", param = list(sigma = 1))
check_full_rank(ef, theta_range = seq(-3, 3, by = 0.5))
}
