% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_confidence_intervals.R
\name{ci_uma}
\alias{ci_uma}
\title{Construct a UMA (Uniformly Most Accurate) Confidence Interval}
\usage{
ci_uma(distribution, parameter, conf_level = 0.95)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{parameter}{Character string: parameter of interest}

\item{conf_level}{Numeric: confidence level}
}
\value{
A message describing what will be computed
}
\description{
Construct a UMA (Uniformly Most Accurate) Confidence Interval
}
