% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_confidence_intervals.R
\name{ci_umau}
\alias{ci_umau}
\title{Construct a UMAU (Uniformly Most Accurate Unbiased) Confidence Interval}
\usage{
ci_umau(distribution, parameter, conf_level = 0.95)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{parameter}{Character string: parameter of interest}

\item{conf_level}{Numeric: confidence level}
}
\value{
A message describing what will be computed
}
\description{
Construct a UMAU (Uniformly Most Accurate Unbiased) Confidence Interval
}
