% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_completeness.R
\name{completeness_basu}
\alias{completeness_basu}
\title{Apply Basu's Theorem}
\usage{
completeness_basu(distribution, suff_stat, anc_stat)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{suff_stat}{Character string: the complete sufficient statistic}

\item{anc_stat}{Character string: the ancillary statistic}
}
\value{
A message describing what will be computed
}
\description{
Apply Basu's Theorem
}
