% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_completeness.R
\name{completeness_family}
\alias{completeness_family}
\title{Check completeness of a family of distributions}
\usage{
completeness_family(family)
}
\arguments{
\item{family}{Character string describing the distributional family}
}
\value{
A message describing what will be computed
}
\description{
Check completeness of a family of distributions
}
