% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_completeness.R
\name{completeness_statistic}
\alias{completeness_statistic}
\title{Check completeness of a statistic T(X)}
\usage{
completeness_statistic(distribution, statistic)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{statistic}{Character string describing the statistic T(x)}
}
\value{
A message describing what will be computed
}
\description{
Check completeness of a statistic T(X)
}
