% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_estimation_methods.R
\name{estimation_minmse}
\alias{estimation_minmse}
\title{Minimum Mean Squared Error (MinMSE) Estimator}
\usage{
estimation_minmse(estimator_class, parameter, distribution)
}
\arguments{
\item{estimator_class}{Character string: class of estimators to search over}

\item{parameter}{Character string: parameter of interest}

\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Minimum Mean Squared Error (MinMSE) Estimator
}
