% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_estimation_methods.R
\name{estimation_mome}
\alias{estimation_mome}
\title{Method of Moments Estimator (MoME)}
\usage{
estimation_mome(distribution, parameters, data = NULL)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{parameters}{Character vector of parameters to estimate}

\item{data}{Optional numeric vector of observed data}
}
\value{
A message describing what will be computed
}
\description{
Method of Moments Estimator (MoME)
}
