% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_exponential_families.R
\name{exp_family_canonical}
\alias{exp_family_canonical}
\title{Express a distribution in canonical (natural) form}
\usage{
exp_family_canonical(distribution)
}
\arguments{
\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Express a distribution in canonical (natural) form
}
