% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_exponential_families.R
\name{exp_family_full_rank}
\alias{exp_family_full_rank}
\title{Determine if an exponential family is full rank}
\usage{
exp_family_full_rank(distribution)
}
\arguments{
\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Determine if an exponential family is full rank
}
