% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_exponential_families.R
\name{exp_family_sufficient_stat}
\alias{exp_family_sufficient_stat}
\title{Extract Sufficient Statistic from Exponential Family}
\usage{
exp_family_sufficient_stat(ef, x)
}
\arguments{
\item{ef}{An "expfam" object}

\item{x}{Numeric vector of observations}
}
\value{
Value of the sufficient statistic T(x)
}
\description{
For an iid sample from an exponential family, returns the (minimal) sufficient statistic.
}
\examples{
ef <- exponential_family("poisson")
x <- rpois(20, lambda = 3)
exp_family_sufficient_stat(ef, x)
}
