% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_expfam_testing.R
\name{expfam_test_neyman_structure}
\alias{expfam_test_neyman_structure}
\title{Tests with Neyman structure in multi-parameter exponential family}
\usage{
expfam_test_neyman_structure(distribution, null_hyp, nuisance, alpha = 0.05)
}
\arguments{
\item{distribution}{Character string: the multi-parameter exponential family}

\item{null_hyp}{Character string: H_0 on the parameter of interest}

\item{nuisance}{Character string: nuisance parameter(s)}

\item{alpha}{Numeric: significance level}
}
\value{
A message describing what will be computed
}
\description{
Tests with Neyman structure in multi-parameter exponential family
}
