% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_information_bounds.R
\name{fisher_information}
\alias{fisher_information}
\title{Fisher Information}
\usage{
fisher_information(
  family = c("normal_mean", "normal_var", "poisson", "binomial", "exponential",
    "gamma_rate", "bernoulli"),
  theta,
  param = list()
)
}
\arguments{
\item{family}{Character: distribution family}

\item{theta}{Numeric: parameter value}

\item{param}{Named list: additional fixed parameters}
}
\value{
Fisher information value
}
\description{
Computes the Fisher information I(theta) = E[{d/dtheta log f(X;theta)}^2]
= -E[d^2/dtheta^2 log f(X;theta)] for common distributions.
}
\examples{
fisher_information("normal_mean", theta = 2, param = list(sigma = 1))
fisher_information("poisson", theta = 3)
fisher_information("binomial", theta = 0.4, param = list(m = 10))
}
