% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_information_bounds.R
\name{fisher_information_matrix}
\alias{fisher_information_matrix}
\title{Fisher Information Matrix (Multiparameter Case)}
\usage{
fisher_information_matrix(
  family = c("normal", "gamma", "beta"),
  theta,
  param = list()
)
}
\arguments{
\item{family}{Character: distribution family}

\item{theta}{Named numeric vector of parameter values}

\item{param}{Additional fixed parameters}
}
\value{
Fisher information matrix
}
\description{
Computes the Fisher information matrix for multiparameter distributions.
I_{jk}(theta) = E[-d^2/dtheta_j dtheta_k log f(X; theta)]
}
\examples{
fisher_information_matrix("normal", theta = c(mu = 2, sigma2 = 4))
fisher_information_matrix("gamma", theta = c(alpha = 2, beta = 1))
}
