% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_world}
\alias{hello_world}
\title{Say Hello}
\usage{
hello_world(name = "World")
}
\arguments{
\item{name}{A character string.}
}
\value{
A greeting message
}
\description{
Prints a greeting message.
}
