% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_umvue.R
\name{hoeffding_u_stat}
\alias{hoeffding_u_stat}
\title{Hoeffding's U-Statistics}
\usage{
hoeffding_u_stat(x, kernel_fn, m = 1, estimand = "theta")
}
\arguments{
\item{x}{Numeric vector of observations}

\item{kernel_fn}{Function of degree m: h(x_1, ..., x_m) symmetric kernel}

\item{m}{Integer: degree of the kernel}

\item{estimand}{Character description of what is being estimated}
}
\value{
U-statistic value with variance estimate
}
\description{
Computes a U-statistic with a given kernel function h for estimating
a population parameter. U-statistics are UMVUE in non-parametric families.
}
\examples{
x <- rnorm(20, mean = 3, sd = 2)

# U-statistic for population mean (m=1)
hoeffding_u_stat(x, kernel_fn = function(x) x, m = 1, estimand = "mu")

# U-statistic for population variance (m=2 kernel)
kernel_var <- function(x1, x2) (x1 - x2)^2 / 2
hoeffding_u_stat(x, kernel_fn = kernel_var, m = 2, estimand = "sigma^2")

# Wilcoxon kernel for P(X > 0) [m=1]
hoeffding_u_stat(x, kernel_fn = function(x) as.numeric(x > 0), m = 1,
                 estimand = "P(X > 0)")
}
