% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_mean.R
\name{infer_mean}
\alias{infer_mean}
\title{Infer Mean using t-test}
\usage{
infer_mean(x, mu = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{mu}{Hypothesized mean}
}
\value{
A list containing test statistic, p-value and confidence interval
}
\description{
Performs a one-sample t-test and returns a tidy result.
}
