% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_information_bounds.R
\name{info_bhattacharya}
\alias{info_bhattacharya}
\title{Compute the Bhattacharya System of Lower Bounds}
\usage{
info_bhattacharya(distribution, parameter, order = 2)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{parameter}{Character string: parameter theta}

\item{order}{Integer: order of the Bhattacharya bound (1 = CRLB, 2, 3, ...)}
}
\value{
A message describing what will be computed
}
\description{
Compute the Bhattacharya System of Lower Bounds
}
