% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_information_bounds.R
\name{info_cramer_rao}
\alias{info_cramer_rao}
\title{Apply the Cramer-Rao Lower Bound}
\usage{
info_cramer_rao(estimator, parameter, distribution)
}
\arguments{
\item{estimator}{Character string: estimator T(X)}

\item{parameter}{Character string: g(theta) being estimated}

\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Apply the Cramer-Rao Lower Bound
}
