% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_information_bounds.R
\name{info_exponential_family}
\alias{info_exponential_family}
\title{Information inequality for s-parameter exponential family}
\usage{
info_exponential_family(distribution, s)
}
\arguments{
\item{distribution}{Character string naming the s-parameter exponential family}

\item{s}{Integer: number of parameters}
}
\value{
A message describing what will be computed
}
\description{
Information inequality for s-parameter exponential family
}
