% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_information_bounds.R
\name{info_fisher}
\alias{info_fisher}
\title{Compute Fisher Information}
\usage{
info_fisher(distribution, parameter)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{parameter}{Character string: the parameter theta}
}
\value{
A message describing what will be computed
}
\description{
Compute Fisher Information
}
