% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_information_bounds.R
\name{info_fisher_matrix}
\alias{info_fisher_matrix}
\title{Compute the Fisher Information Matrix (multi-parameter)}
\usage{
info_fisher_matrix(distribution, parameters)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{parameters}{Character vector of parameter names}
}
\value{
A message describing what will be computed
}
\description{
Compute the Fisher Information Matrix (multi-parameter)
}
