% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_information_bounds.R
\name{info_hcr_bound}
\alias{info_hcr_bound}
\title{Apply the Hammersley-Chapman-Robbins (HCR) Inequality}
\usage{
info_hcr_bound(estimator, parameter, distribution)
}
\arguments{
\item{estimator}{Character string: estimator T(X)}

\item{parameter}{Character string: g(theta) being estimated}

\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Apply the Hammersley-Chapman-Robbins (HCR) Inequality
}
