% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_sufficiency.R
\name{sufficiency_ancillary}
\alias{sufficiency_ancillary}
\title{Check if a statistic is ancillary}
\usage{
sufficiency_ancillary(distribution, statistic)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{statistic}{Character string describing the candidate statistic}
}
\value{
A message describing what will be computed
}
\description{
Check if a statistic is ancillary
}
