% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_sufficiency.R
\name{sufficiency_minimal}
\alias{sufficiency_minimal}
\title{Check minimal sufficiency of a statistic}
\usage{
sufficiency_minimal(distribution, statistic)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{statistic}{Character string describing the candidate statistic T(x)}
}
\value{
A message describing what will be computed
}
\description{
Check minimal sufficiency of a statistic
}
