% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_sufficiency.R
\name{sufficiency_nf_factorisation}
\alias{sufficiency_nf_factorisation}
\title{Apply the Neyman-Fisher Factorisation Criterion}
\usage{
sufficiency_nf_factorisation(distribution, statistic)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{statistic}{Character string describing the candidate statistic T(x)}
}
\value{
A message describing what will be computed
}
\description{
Apply the Neyman-Fisher Factorisation Criterion
}
