% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_hypothesis_basics.R
\name{test_critical_region}
\alias{test_critical_region}
\title{Analyse the critical region of a test}
\usage{
test_critical_region(test_statistic, critical_region, distribution)
}
\arguments{
\item{test_statistic}{Character string: the test statistic T(X)}

\item{critical_region}{Character string: description of the rejection region C}

\item{distribution}{Character string naming the distribution under H_0}
}
\value{
A message describing what will be computed
}
\description{
Analyse the critical region of a test
}
