% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_hypothesis_basics.R
\name{test_define_hypothesis}
\alias{test_define_hypothesis}
\title{Define and classify a statistical hypothesis}
\usage{
test_define_hypothesis(null_hyp, alt_hyp)
}
\arguments{
\item{null_hyp}{Character string: the null hypothesis H_0}

\item{alt_hyp}{Character string: the alternative hypothesis H_1}
}
\value{
A message describing what will be computed
}
\description{
Define and classify a statistical hypothesis
}
