% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_hypothesis_basics.R
\name{test_errors_and_power}
\alias{test_errors_and_power}
\title{Compute Type I error, Type II error, size and power}
\usage{
test_errors_and_power(
  null_hyp,
  alt_hyp,
  test_statistic,
  critical_region,
  distribution
)
}
\arguments{
\item{null_hyp}{Character string: H_0}

\item{alt_hyp}{Character string: H_1}

\item{test_statistic}{Character string: T(X)}

\item{critical_region}{Character string: rejection region}

\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Compute Type I error, Type II error, size and power
}
