% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_expfam_testing.R
\name{test_likelihood_ratio}
\alias{test_likelihood_ratio}
\title{Likelihood Ratio Test (LRT)}
\usage{
test_likelihood_ratio(distribution, null_hyp, alt_hyp, alpha = 0.05)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{null_hyp}{Character string: H_0 (can be composite)}

\item{alt_hyp}{Character string: H_1}

\item{alpha}{Numeric: significance level}
}
\value{
A message describing what will be computed
}
\description{
Likelihood Ratio Test (LRT)
}
