% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_np_lemma_tests.R
\name{test_mlr}
\alias{test_mlr}
\title{Test using Monotone Likelihood Ratio (MLR) property}
\usage{
test_mlr(distribution, statistic, null_hyp, alt_hyp, alpha = 0.05)
}
\arguments{
\item{distribution}{Character string naming the distribution}

\item{statistic}{Character string: the statistic with MLR property}

\item{null_hyp}{Character string: H_0}

\item{alt_hyp}{Character string: H_1}

\item{alpha}{Numeric: significance level}
}
\value{
A message describing what will be computed
}
\description{
Test using Monotone Likelihood Ratio (MLR) property
}
