% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_np_lemma_tests.R
\name{test_mp}
\alias{test_mp}
\title{Construct a Most Powerful (MP) test}
\usage{
test_mp(null_dist, alt_dist, alpha = 0.05)
}
\arguments{
\item{null_dist}{Character string: distribution under H_0 (simple)}

\item{alt_dist}{Character string: distribution under H_1 (simple)}

\item{alpha}{Numeric: significance level}
}
\value{
A message describing what will be computed
}
\description{
Construct a Most Powerful (MP) test
}
