% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_np_lemma_tests.R
\name{test_umpu}
\alias{test_umpu}
\title{Construct an Unbiased Test and UMPU Test}
\usage{
test_umpu(null_hyp, alt_hyp, distribution, alpha = 0.05)
}
\arguments{
\item{null_hyp}{Character string: H_0}

\item{alt_hyp}{Character string: H_1}

\item{distribution}{Character string naming the distributional family}

\item{alpha}{Numeric: significance level}
}
\value{
A message describing what will be computed
}
\description{
Construct an Unbiased Test and UMPU Test
}
