% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_unbiasedness_umvue.R
\name{umvue_find}
\alias{umvue_find}
\title{Find the Uniformly Minimum Variance Unbiased Estimator (UMVUE)}
\usage{
umvue_find(parameter, distribution)
}
\arguments{
\item{parameter}{Character string: parameter of interest (e.g. "theta", "theta^2")}

\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Find the Uniformly Minimum Variance Unbiased Estimator (UMVUE)
}
