% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_unbiasedness_umvue.R
\name{umvue_lehmann_scheffe}
\alias{umvue_lehmann_scheffe}
\title{Apply Lehmann-Scheffe's Theorem}
\usage{
umvue_lehmann_scheffe(estimator, suff_stat, distribution)
}
\arguments{
\item{estimator}{Character string: unbiased estimator that is a function of T}

\item{suff_stat}{Character string: complete sufficient statistic T(X)}

\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Apply Lehmann-Scheffe's Theorem
}
