% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_umvue_methods.R
\name{umvue_method_direct}
\alias{umvue_method_direct}
\title{Find UMVUE by direct method (solving integral equations)}
\usage{
umvue_method_direct(parameter, suff_stat, distribution)
}
\arguments{
\item{parameter}{Character string: g(theta) to estimate}

\item{suff_stat}{Character string: complete sufficient statistic}

\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Find UMVUE by direct method (solving integral equations)
}
