% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_umvue_methods.R
\name{umvue_method_rao_blackwell}
\alias{umvue_method_rao_blackwell}
\title{Find UMVUE by Rao-Blackwellisation}
\usage{
umvue_method_rao_blackwell(raw_estimator, suff_stat, distribution)
}
\arguments{
\item{raw_estimator}{Character string: a simple unbiased estimator to start with}

\item{suff_stat}{Character string: complete sufficient statistic T(X)}

\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Find UMVUE by Rao-Blackwellisation
}
