% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_umvue.R
\name{umvue_normal_mean}
\alias{umvue_normal_mean}
\title{UMVUE for Normal Mean (sigma known)}
\usage{
umvue_normal_mean(x, sigma = 1)
}
\arguments{
\item{x}{Numeric vector of observations}

\item{sigma}{Known standard deviation}
}
\value{
List with UMVUE estimate and properties
}
\description{
The UMVUE of mu for X_i ~ N(mu, sigma^2) is X_bar (sample mean).
By Lehmann-Scheffe: X_bar is a function of the complete sufficient statistic
and is unbiased, hence it is the UMVUE.
}
\examples{
x <- rnorm(25, mean = 3, sd = 2)
umvue_normal_mean(x, sigma = 2)
}
