% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_umvue_methods.R
\name{umvue_ustatistic}
\alias{umvue_ustatistic}
\title{Construct Hoeffding's U-statistic for non-parametric families}
\usage{
umvue_ustatistic(kernel, degree, parameter, n)
}
\arguments{
\item{kernel}{Character string: the symmetric kernel h(x_1, ..., x_m)}

\item{degree}{Integer: degree m of the kernel}

\item{parameter}{Character string: the functional theta being estimated}

\item{n}{Integer: sample size}
}
\value{
A message describing what will be computed
}
\description{
Construct Hoeffding's U-statistic for non-parametric families
}
