% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_unbiasedness_umvue.R
\name{unbiased_check}
\alias{unbiased_check}
\title{Check basic unbiasedness of an estimator}
\usage{
unbiased_check(estimator, parameter, distribution)
}
\arguments{
\item{estimator}{Character string: estimator T(X)}

\item{parameter}{Character string: parameter being estimated (e.g. "mu", "sigma^2")}

\item{distribution}{Character string naming the distribution}
}
\value{
A message describing what will be computed
}
\description{
Check basic unbiasedness of an estimator
}
