% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_np_lemma_tests.R
\name{np_lemma}
\alias{np_lemma}
\title{Apply the Neyman-Pearson Lemma}
\usage{
np_lemma(null_dist, alt_dist, alpha = 0.05)
}
\arguments{
\item{null_dist}{Character string: distribution under H_0}

\item{alt_dist}{Character string: distribution under H_1}

\item{alpha}{Numeric: significance level}
}
\value{
A message describing what will be computed
}
\description{
Apply the Neyman-Pearson Lemma
}
